feat(props$): useEpics options.props will now do a shallow diff to see if it emits props
Passing objects to the `options.props` property will now be shallow diffed and only if a value is
different will the new props be emitted.  This will work better for just passing your entire props
object into the options.props property

BREAKING CHANGE: If anyone was relying on just the object being a different reference, this will now
break. the useEpic options.props property will be shallowly diffed and if there are now changes to
the property values, new props will not be emiitted to the props$ Observable in the epic deps
argument